-Finish login switcher
